requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.20.0
accelerate>=0.20.0

# Image processing
Pillow>=9.5.0
opencv-python>=4.8.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web interface
streamlit>=1.25.0

# Configuration and utilities
PyYAML>=6.0
requests>=2.31.0
tqdm>=4.65.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: For better performance
# xformers>=0.0.20  # Uncomment for better memory efficiency
# bitsandbytes>=0.41.0  # Uncomment for 8-bit optimization
